make auth separate pages
