+++ microblog/src/routes/auth/logout/+page.server.ts
@@ -1,0 +1,21 @@
+import { fail, redirect } from "@sveltejs/kit";
+import type { Actions } from "./$types";
+import { lucia } from "$lib/server/auth";
+
+export const actions: Actions = {
+	default: async (event) => {
+		// console.log('logout')
+		if (!event.locals.session) {
+			return fail(401);
+		}
+
+		await lucia.invalidateSession(event.locals.session.id);
+		const sessionCookie = lucia.createBlankSessionCookie();
+		event.cookies.set(sessionCookie.name, sessionCookie.value, {
+			path: '.',
+			...sessionCookie.attributes
+		});
+
+		return redirect(302, '/auth/redirect-root')
+	}
+}
